+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 * moved the method `extend_or_make_path()` from the `xx_file` module to the `xx_path` module and renamed it to `extend_or_make()`
 * added a new param to method `Color.luminance()` and to the `.grayscale()` method of all color types:
   - <code>method: *str* = "wcag2"</code> the luminance calculation method to use
+* added a new param to the method `File.rename_extension()`:
+  - <code>full_extension: *bool* = False</code> whether to treat everything behind the first `.` as the extension or everything behind the last `.`
 
 ## 18.03.2025 `v1.6.8`
 * made it possible to escape formatting codes by putting a slash (`/` *or* `\\`) at the beginning inside the brackets (*e.g.* `[/red]`)